requirements.txt
streamlit
python-dotenv
openai
tiktoken
PyPDF2
python-pptx
tqdm
pdfplumber
numpy
pandas
